+++ homeassistant/components/knx/config_flow.py
@@ -57,7 +57,7 @@ from .const import (
     KNXConfigEntryData,
 )
 from .helpers.keyring import DEFAULT_KNX_KEYRING_FILENAME, save_uploaded_knxkeys_file
-from .schema import ia_validator, ip_v4_validator
+from .validation import ia_validator, ip_v4_validator
 
 CONF_KNX_GATEWAY: Final = "gateway"
 CONF_MAX_RATE_LIMIT: Final = 60
